+++ src/shared/page.ts
@@ -243,6 +243,56 @@ export default class Page<api = {}> extends LitElement {
       flex: 1;
       max-width: 150px;
     }
+    .switch {
+      position: relative;
+      display: inline-block;
+      width: 48px;
+      height: 24px;
+      flex-shrink: 0;
+      margin-top: 7px;
+    }
+
+    .switch input {
+      opacity: 0;
+      width: 0;
+      height: 0;
+    }
+
+    .slider {
+      position: absolute;
+      cursor: pointer;
+      top: 0;
+      left: 0;
+      right: 0;
+      bottom: 0;
+      background-color: #ccc;
+      transition: .4s;
+      border-radius: 24px;
+    }
+
+    .slider:before {
+      position: absolute;
+      content: "";
+      height: 18px;
+      width: 18px;
+      left: 3px;
+      bottom: 3px;
+      background-color: white;
+      transition: .4s;
+      border-radius: 50%;
+    }
+
+    input:checked + .slider {
+      background-color: var(--group-button-active-bg);
+    }
+
+    input:focus + .slider {
+      box-shadow: 0 0 1px var(--group-button-active-bg);
+    }
+
+    input:checked + .slider:before {
+      transform: translateX(24px);
+    }
     ` as CSSResultGroup];
 
   /**