+++ src/components/componentUserStatus/CHANGELOG.md
@@ -1,5 +1,11 @@
 # component-user-status
 
+## 0.0.2
+
+### Patch Changes
+
+- adding periodical notification reminder to fill daily user status data
+
 ## 0.0.1
 
 ### Patch Changes