+++ src/components/pageHome/CHANGELOG.md
@@ -1,5 +1,11 @@
 # home-page
 
+## 0.0.12
+
+### Patch Changes
+
+- adding periodical notification reminder to fill daily user status data
+
 ## 0.0.11
 
 ### Patch Changes